+++ src/hooks/index.ts
@@ -3,3 +3,4 @@ export { useThrottle } from './useThrottle';
 export { useLocalStorage } from './useLocalStorage';
 export { useSessionStorage } from './useSessionStorage';
 export { useCookie } from './useCookie';
+export { useClickOutside } from './useClickOutside';